% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test.R
\name{compute_test}
\alias{compute_test}
\title{Run random forest models to find significant predictive relationships}
\usage{
compute_test(data_list)
}
\arguments{
\item{data_list}{A list containing data and its attributes (preferably from data_reshape())}
}
\value{
A list containing results of random forest model (MDG, p-value, optimized number of trees, etc)
}
\description{
This funciton organizes the data and computes mean decrease in Gini (MDG) and associated p-values.
}
\examples{

#' @importFrom ranger ranger
ia_test2 <- compute_test(ia_test1)
}
