% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tuning}
\alias{tuning}
\title{Tuning random forest}
\usage{
tuning(data, target, num_permutation)
}
\arguments{
\item{data}{A dataframe containing possible predictors from same and preceding blocks}

\item{target}{A character representing response variable for random forest model}

\item{num_permutation}{A numeric number indicating number of permutation for Altmann's approach}
}
\value{
A list containing results of tuning process
}
\description{
Tuning random forest
}
