% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Network visualization for predictie relationships}
\usage{
visualize(data, edge_width = 2, cutoff = 0.05, shape = "dot", size = 12, color)
}
\arguments{
\item{data}{A list to visualize the results (preferably from compute_test())}

\item{edge_width}{Width of edge in the graph}

\item{cutoff}{A cutoff for significance}

\item{shape}{A character for shape of nodes in the graph}

\item{size}{Node and text size for graph and its legend}

\item{color}{A character vector to distinguish each year in the data (length: number of measurements)}
}
\description{
Network visualization for predictie relationships
}
\examples{
color_brewer <- RColorBrewer::brewer.pal(8, "Pastel1")

visualize(ia_test2, color = color_brewer[1:3])
}
