% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test.R
\name{compute_test}
\alias{compute_test}
\title{Run random forest models to find significant predictive relationships}
\usage{
compute_test(data_list)
}
\arguments{
\item{data_list}{A list containing data and its attributes (preferably from data_reshape())}
}
\value{
A list containing results of random forest model (MDG, p-value, optimized number of trees, etc)
}
\description{
This funciton organizes the data and computes mean decrease in Gini (MDG) and associated p-values.
}
\examples{

#' @importFrom ranger ranger
ia_test2 <- compute_test(ia_test1)
}
\references{
\itemize{
\item Breiman, L. (2001). Random forests. Mach Learn, 45:5-32. https://doi.org/10.1023/A:1010933404324.
\item Wright, M. N. & Ziegler, A. (2017). ranger: A fast implementation of random forests for high dimensional data in C++ and R. J Stat Softw 77:1-17. https://doi.org/10.18637/jss.v077.i01.
\item Altmann A, Toloşi L, Sander O, Lengauer T. Permutation importance: a corrected feature importance measure. Bioinformatics. 2010;26(10):1340-1347.
}
}
